Unify code with signatureFragments & normalizedHash functions in MoleculeStructure class
